requirements.txt
unstructured
tiktoken
langchain
streamlit
pinecone-client
pypdf
streamlit_chat 
sentence_transformers
openai